made pid removal in finally dependant on the existence of the pid file

+++ extlib.bash
@@ -69,7 +69,10 @@ quit () {
 
 finally () {
     # Function to perform final tasks before exit
-    rm "${PIDFILE}"
+    if [ -f "${PIDFILE}" ]
+    then
+        rm "${PIDFILE}"
+    fi
 }
 
 # Trap to do final tasks before exit
